error handling implemented for the login form
the error handling appears different to the signup form, perhaps there is a better way to handle this/nicer way to handle it. Doing it similar to the signup form seems to give no error output, whereas taking the form dictionary and narrowing it down to the errors appears to work. Seems overly complicated/un-necessary, but if it's the only way then so be it